Changed the location of object detection folder
