+++ src/Football-object-detection/requirements.txt
@@ -1,0 +1,4 @@
+numpy==1.22.4
+opencv_python==4.6.0.66
+scikit_learn==1.3.0
+ultralytics==8.0.168
