+++ src/styles.css
@@ -1,0 +1,64 @@
+body {
+    font-family: Arial, sans-serif;
+    background-color: #f8f9fa;
+    margin: 0;
+    padding: 0;
+}
+
+h1 {
+    color: #2C3E50;
+    text-align: center;
+    padding-top: 20px;
+}
+
+.form-container {
+    text-align: center;
+    margin-top: 20px;
+}
+
+.input, .button {
+    font-size: 1rem;
+    padding: 10px;
+    border-radius: 5px;
+    margin: 5px;
+}
+
+.container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 20px;
+    margin-top: 20px;
+}
+
+.video-box {
+    width: 48%;
+    position: relative;
+}
+
+video {
+    width: 100%;
+    border-radius: 10px;
+    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.2);
+}
+
+.controls {
+    display: flex;
+    justify-content: center;
+    margin-top: 10px;
+}
+
+.button {
+    background-color: #2C3E50;
+    color: white;
+    border: none;
+    cursor: pointer;
+    padding: 10px 20px;
+    border-radius: 5px;
+    font-size: 1.2rem;
+}
+
+.button:hover {
+    background-color: #1A2A38;
+}
+
